Attempts to re-establish dropped connections, creates outbound connections for inbound connections if they don't exist yet. Created ServerUtilities class which contains many messaging functions and contains the function to connect servers. Added a 3rd server for testing
