+++ server-files/server_list.h
@@ -14,7 +14,7 @@ class ServerList{
         std::unordered_map<int, std::string> knownClients; // Clients that belong to this server
 
         // Temporary way to store server addresses against their ID
-        std::unordered_map<int, std::string> serverAddresses = {{1, "127.0.0.1:9002"}, {2, "127.0.0.1:9003"}};
+        std::unordered_map<int, std::string> serverAddresses = {{1, "127.0.0.1:9002"}, {2, "127.0.0.1:9003"}, {3, "127.0.0.1:9004"}};
 
         void save_mapping_to_file();
         void load_mapping_from_file();
@@ -25,6 +25,7 @@ class ServerList{
         ServerList(int server_id);
 
         int ObtainID(std::string address);
+        std::unordered_map<int, std::string> getUris();
 
         std::pair<int, std::string> retrieveClient(int server_id, int client_id);
