+++ server-files/server_mapping3.json
@@ -1,0 +1,6 @@
+[
+    [
+        1001,
+        "GHIJKL"
+    ]
+]
